requirements.txt
alembic           1.13.1
aniso8601         9.0.1
blinker           1.8.2
click             8.1.7
colorama          0.4.6
Flask             3.0.3
Flask-Migrate     4.0.7
Flask-RESTful     0.3.10
Flask-Script      2.0.6
Flask-SQLAlchemy  3.1.1
Flask-WTF         1.2.1
greenlet          3.0.3
itsdangerous      2.2.0
Jinja2            3.1.4
Mako              1.3.5
MarkupSafe        2.1.5
pip               23.2.1
psycopg           3.1.19
psycopg2-binary   2.9.9
pytz              2024.1
six               1.16.0
SQLAlchemy        2.0.30
typing_extensions 4.12.1
tzdata            2024.1
Werkzeug          3.0.3
WTForms           3.1.2
